Pre-loop working for generic PLC. Next step should be generalizing the main method
Pre-loop working for generic PLC. Next step should be generalizing the main method

First version of generic_automatic_plc an plc classes. We need to do additional testing, but so far the plc seems to behave correctly by the plc_dict genearted by the epanet parser. If this works, all that's left is to generalize the automatic_run.py and topo.py behaviour

Added a CONTROL flag in the db to sync the actuator state and physical process state

Generic plc still not working

Generic plc still not working

First version of generic_automatic_plc an plc classes. We need to do additional testing, but so far the plc seems to behave correctly by the plc_dict genearted by the epanet parser. If this works, all that's left is to generalize the automatic_run.py and topo.py behaviour

We have a working version of generic_plc. This class supports any control logic defined in EPANET files when an actuator control decision depends on a ON/OFF logic with only one variable. Some EPANET topologies work in an open loop way, those should be handled by the physical_process.py

Right now, we are working in generalizing the topologies and automatic_run. Looking good so far, topo files will need plc_dicts to know how many PLC nodes need to be launched. general_automatic_run.py already can handle launching either simple or complex topo.

general_automatic_run.py should have a method that checks into the PLC dependencies to decide launch and order, after we implement all this, DHALSIM should be general for the EPANET topologies. Next should be working in an integrated framework attack

general_automatic_run has been generalized. Nevertheless, we still need to do some work. i) The finish process is failing, we get an index out of bounds exception, ii) We gotta rework in the epanet_parser.py, because we need to add the capability of generating an utils file for simple or complex network topologies. In a similar way, we need to finish the code in simple_topo.py
After of this, there is a lot of testing to do

general_automatic_run has been generalized. Nevertheless, we still need to do some work. i) We gotta rework in the epanet_parser.py, because we need to add the capability of generating an utils file for simple or complex network topologies. In a similar way, we need to finish the code in simple_topo.py
After of this, there is a lot of testing to do

general_automatic_run has been generalized. Nevertheless, we still need to do some work. i) We gotta rework in the epanet_parser.py, because we need to add the capability of generating an utils file for simple or complex network topologies. In a similar way, we need to finish the code in simple_topo.py
After of this, there is a lot of testing to do

Added ky3 topo to test the general_automatic_run

Added ky3 topo to test the general_automatic_run

Added ky3 topo to test the general_automatic_run

Added ky3 topo to test the general_automatic_run

Pre-loop working for generic PLC. Next step should be generalizing the main method

First version of generic_automatic_plc an plc classes. We need to do additional testing, but so far the plc seems to behave correctly by the plc_dict genearted by the epanet parser. If this works, all that's left is to generalize the automatic_run.py and topo.py behaviour

Added a CONTROL flag in the db to sync the actuator state and physical process state

Generic plc still not working

First version of generic_automatic_plc an plc classes. We need to do additional testing, but so far the plc seems to behave correctly by the plc_dict genearted by the epanet parser. If this works, all that's left is to generalize the automatic_run.py and topo.py behaviour

We have a working version of generic_plc. This class supports any control logic defined in EPANET files when an actuator control decision depends on a ON/OFF logic with only one variable. Some EPANET topologies work in an open loop way, those should be handled by the physical_process.py

Right now, we are working in generalizing the topologies and automatic_run. Looking good so far, topo files will need plc_dicts to know how many PLC nodes need to be launched. general_automatic_run.py already can handle launching either simple or complex topo.

general_automatic_run.py should have a method that checks into the PLC dependencies to decide launch and order, after we implement all this, DHALSIM should be general for the EPANET topologies. Next should be working in an integrated framework attack

general_automatic_run has been generalized. Nevertheless, we still need to do some work. i) The finish process is failing, we get an index out of bounds exception, ii) We gotta rework in the epanet_parser.py, because we need to add the capability of generating an utils file for simple or complex network topologies. In a similar way, we need to finish the code in simple_topo.py
After of this, there is a lot of testing to do

general_automatic_run has been generalized. Nevertheless, we still need to do some work. i) We gotta rework in the epanet_parser.py, because we need to add the capability of generating an utils file for simple or complex network topologies. In a similar way, we need to finish the code in simple_topo.py
After of this, there is a lot of testing to do

Added ky3 topo to test the general_automatic_run

Added ky3 topo to test the general_automatic_run

Added ky3 topo to test the general_automatic_run