Added a jupyter notebook to compare getting results from the sim.results object or from waternetworkmodel node objects. Also, we wil use thie notebook to compare EPANET and WNTR simulators
Added temporary .csv results file to debug the EPANET simulator integration.
So far, EPANET works in standalone mode with similar results to the WNTR simulator (albeit faster). Nevertheless, trying to use the EPANET simulator with MiniCPS generates an EPANET 200 error during the actuator_update method of physical_process.py. Despite the error, the actuator syntax of the water network model from standalone and MiniCPS looks identical

Minor bug fixes

Fixed an issue that was not allowing us to use the EPANET simulator in WNTR! Now the simulation runs faster. The problem was caused by the difference between mininum and required pressure

Fixed an issue that was not allowing us to use the EPANET simulator in WNTR! Now the simulation runs faster. The problem was caused by the difference between mininum and required pressure

Fixed an issue that was not allowing us to use the EPANET simulator in WNTR! Now the simulation runs faster. The problem was caused by the difference between mininum and required pressure

Added temporary .csv results file to debug the EPANET simulator integration.
So far, EPANET works in standalone mode with similar results to the WNTR simulator (albeit faster). Nevertheless, trying to use the EPANET simulator with MiniCPS generates an EPANET 200 error during the actuator_update method of physical_process.py. Despite the error, the actuator syntax of the water network model from standalone and MiniCPS looks identical

Changes to the enhanced_ctown_topology to use the newly added method to the EPANET simulator

Added temporary .csv results file to debug the EPANET simulator integration.
So far, EPANET works in standalone mode with similar results to the WNTR simulator (albeit faster). Nevertheless, trying to use the EPANET simulator with MiniCPS generates an EPANET 200 error during the actuator_update method of physical_process.py. Despite the error, the actuator syntax of the water network model from standalone and MiniCPS looks identical

Added temporary .csv results file to debug the EPANET simulator integration.
So far, EPANET works in standalone mode with similar results to the WNTR simulator (albeit faster). Nevertheless, trying to use the EPANET simulator with MiniCPS generates an EPANET 200 error during the actuator_update method of physical_process.py. Despite the error, the actuator syntax of the water network model from standalone and MiniCPS looks identical

Added temporary .csv results file to debug the EPANET simulator integration.
So far, EPANET works in standalone mode with similar results to the WNTR simulator (albeit faster). Nevertheless, trying to use the EPANET simulator with MiniCPS generates an EPANET 200 error during the actuator_update method of physical_process.py. Despite the error, the actuator syntax of the water network model from standalone and MiniCPS looks identical

Added temporary .csv results file to debug the EPANET simulator integration.
So far, EPANET works in standalone mode with similar results to the WNTR simulator (albeit faster). Nevertheless, trying to use the EPANET simulator with MiniCPS generates an EPANET 200 error during the actuator_update method of physical_process.py. Despite the error, the actuator syntax of the water network model from standalone and MiniCPS looks identical

Added temporary .csv results file to debug the EPANET simulator integration.
So far, EPANET works in standalone mode with similar results to the WNTR simulator (albeit faster). Nevertheless, trying to use the EPANET simulator with MiniCPS generates an EPANET 200 error during the actuator_update method of physical_process.py. Despite the error, the actuator syntax of the water network model from standalone and MiniCPS looks identical

Changes to the enhanced_ctown_topology to use the newly added method to the EPANET simulator

Tcpdump receives now a SIGINT instead of a SIGKILL. Now our pcap files do not end in the middle of a capture

Tcpdump receives now a SIGINT instead of a SIGKILL. Now our pcap files do not end in the middle of a capture.
Changed the moving to the week_x folder from basePLC to automatic_run, makes more sense. Also, added a step in the shutdown process to merge the pcap files using tshark, we sould modify the documentation to add that dependency

Tcpdump receives now a SIGINT instead of a SIGKILL. Now our pcap files do not end in the middle of a capture.
Changed the moving to the week_x folder from basePLC to automatic_run, makes more sense. Also, added a step in the shutdown process to merge the pcap files using tshark, we sould modify the documentation to add that dependency

Adding junctions pressure and pump flow to the PLCs and SCADA. These data is going to be received by the SCADA server. This is fundamental to implement attacks with concealment

Tcpdump receives now a SIGINT instead of a SIGKILL. Now our pcap files do not end in the middle of a capture.
Changed the moving to the week_x folder from basePLC to automatic_run, makes more sense. Also, added a step in the shutdown process to merge the pcap files using tshark, we sould modify the documentation to add that dependency

Tcpdump receives now a SIGINT instead of a SIGKILL. Now our pcap files do not end in the middle of a capture.
Changed the moving to the week_x folder from basePLC to automatic_run, makes more sense. Also, added a step in the shutdown process to merge the pcap files using tshark, we sould modify the documentation to add that dependency