Finishing the merge into master. Although the EPANET simulator still cannot be used, we fixed a lot of minor issues with previous topologies
Finishing the merge into master.
Although the EPANET simulator still cannot be used, we fixed a lot of minor issues with previous topologies

Adding a file that had caused conflict

Finishing the merge into master.
Although the EPANET simulator still cannot be used, we fixed a lot of minor issues with previous topologies

Fixed a bug into BasePLC. The function to send tags values was sending duplicate values

Added a couple of jupyter notebooks to help handling data

Added a couple of jupyter notebooks to help handling data

Added a couple of jupyter notebooks to help handling data

Added a couple of jupyter notebooks to help handling data

Changes to the enhanced_ctown_topology to use the newly added method to the EPANET simulator

Implementing the new attacks on C-Town

Added a couple of jupyter notebooks to help handling data

Implementing the new attacks on C-Town

Implementing the new attacks on C-Town

Started the pipeline to train the Autoencoder

Added attack flags to the SCADA data

Started the pipeline to train the Autoencoder

Started the pipeline to train the Autoencoder

Started the pipeline to train the Autoencoder

Implemented ky3 topology with PLC1 attack

Changes to topo. We now assign "manually" MAC addresses to the routers interfaces

Started the pipeline to train the Autoencoder

Now MAC addresses are configured explicity in the topo.py of enhanced ctown. This was because Mininet's random MAC addresses were affecting the training, testing, and validation datasets. Although we used the setAutoMacs option to true, we needed to add them for the routers, since their class node is different. There is still an issue with the command that parses .pcap files into .csv files. We can run that manually afterwards for now

Adding a file that had caused conflict

Now MAC addresses are configured explicity in the topo.py of enhanced ctown. This was because Mininet's random MAC addresses were affecting the training, testing, and validation datasets. Although we used the setAutoMacs option to true, we needed to add them for the routers, since their class node is different. There is still an issue with the command that parses .pcap files into .csv files. We can run that manually afterwards for now

Changes to topo. We now assign "manually" MAC addresses to the routers interfaces