Initializing all the RNG with the same seed
Testing Davide's version of skipping intermediate steps.

Testing Davide's version of skipping intermediate steps.

Testing Davide's version of skipping intermediate steps.

Testing Davide's version of skipping intermediate steps.

Initializing all the RNG with the same seed

test to fix outdate data

Initializing all the RNG with the same seed

Fixed the synchronization between PLCs and physical process. Updated the simulation to register the initial values of the simulation at iteration 0 using epynet. We still need to confirm the values for the junctions pressure and pumps/valves flow. In addition, we need to extend the synchronization mechanism to the network attack scripts and network events

Fixed the synchronization between PLCs and physical process. Updated the simulation to register the initial values of the simulation at iteration 0 using epynet. We still need to confirm the values for the junctions pressure and pumps/valves flow. In addition, we need to extend the synchronization mechanism to the network attack scripts and network events

Fixed the synchronization between PLCs and physical process. Updated the simulation to register the initial values of the simulation at iteration 0 using epynet. We still need to confirm the values for the junctions pressure and pumps/valves flow. In addition, we need to extend the synchronization mechanism to the network attack scripts and network events

Fixed the synchronization between PLCs and physical process. Updated the simulation to register the initial values of the simulation at iteration 0 using epynet. We still need to confirm the values for the junctions pressure and pumps/valves flow. In addition, we need to extend the synchronization mechanism to the network attack scripts and network events

Fixed the synchronization between PLCs and physical process. Updated the simulation to register the initial values of the simulation at iteration 0 using epynet. We still need to confirm the values for the junctions pressure and pumps/valves flow. In addition, we need to extend the synchronization mechanism to the network attack scripts and network events

Applied the same synchronization scheme to network events

Applied the same synchronization scheme to network attacks

Applied the same synchronization scheme to network attacks

Applied the same synchronization scheme to network attacks

Applied the same synchronization scheme to network attacks

Applied the same synchronization scheme to network attacks

Applied the same synchronization scheme to network attacks

Applied the same synchronization scheme to network attacks

Minor changes to installation scripts, because setuptools 60.0 is causing installation issues