Testing the CIP parser
Working with the parser

Very small fix into register_initial results for initial junction state using wntr simulator

Testing the CIP parser in the concealment_mitm. For now, we have a mechanism that can filter the tags and only modify the tag configured in the attack YAML file. We should move this mechanism to naive_mitm before continuing with the concealment attack.

Testing the CIP parser

Testing the CIP parser

Fix in the gassuian sensor noise formula, to make it a percentage of the measured value

Testing the CIP parser

Testing the CIP parser

Testing the CIP parser

Fix in the gassuian sensor noise formula, to make it a percentage of the measured value

Working on the MiTM concealment attack

Working on the MiTM concealment attack

Testing the CIP parser

Attempting less update retries